requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-rest-passwordreset
stripe
openai==1.16.1
celery==5.4.0
httpx==0.27.2